requirements.txt
accelerate>=1.0.0
gradio==5.39.0
librosa==0.11.0
soundfile==0.13.1
numpy==2.2.6
requests==2.32.4
datasets>=2.20
transformers==4.55.1
triton-windows==3.3.1.post19
bitsandbytes==0.46.1
beautifulsoup4==4.13.4
tinytag==2.1.1
pytorch_lightning==2.5.1
loguru==0.7.3
num2words==0.5.14
spacy==3.8.4
peft>=0.11
prodigyopt>=1.0
py3langid==0.3.0
pypinyin==0.53.0
pandas==2.3.1
tqdm==4.67.1
nltk==3.9.1
matplotlib==3.10.5
huggingface-hub==0.34.3
tokenizers==0.21.4
h5py==3.13.0
hdf5plugin==5.1.0
diffusers>=0.33.0
natsort==8.4.0
rapidfuzz>=3.0
uvicorn>=0.29
tensorboard
tensorboardX